backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
python-multipart

# Rate limiting
slowapi
redis

# Google Gemini (Gen AI)
google-genai

# LlamaIndex (keeping core for potential future usage, though simplified now)
llama-index-core
